Candidate details changes

+++ src/pages/candidateinfo/candidateinfocomponents/Employement.tsx
@@ -21,6 +21,7 @@ import { DataTable } from 'primereact/datatable'
 import { Column } from 'primereact/column'
 import { InputTextarea } from 'primereact/inputtextarea'
 import { Tooltip } from 'primereact/tooltip';
+import { ConfirmDialog, confirmDialog } from 'primereact/confirmdialog'
 
 
 function Employement() {
@@ -118,9 +119,9 @@ function Employement() {
       return (
         <div style={{ overflowY: "scroll", height: "70px" }}>{
           rowdata.files.map((f) => (
-            <div className='field row-12 md:row-12 flex' onClick={() => console.log(f.file)} style={{ border: "2px solid blue", padding: "4px", margin: "4px", borderRadius: "10px", backgroundColor: "#C1C2F3", height: "30px" }}>
+            <div className='field row-12 md:row-12 flex' onClick={() => console.log(f.file)} style={{ border: "2px solid blue", color:"blue", padding: "4px", margin: "4px", borderRadius: "10px", backgroundColor: "#C1C2F3", height: "30px" }}>
 
-              {f.file.split("/")[f.file.split("/").length - 1].toString().length < 20 ? f.file.split("/")[f.file.split("/").length - 1] : f.file.split("/")[f.file.split("/").length - 1].substring(0, 20) + "..."}
+              {f.file.split("/")[f.file.split("/").length - 1].toString().length < 20 ? f.file.split("/")[f.file.split("/").length - 1] : f.file.split("/")[f.file.split("/").length - 1]}
               <i className="pi pi-download mr-2 ml-2" onClick={() => {
                 dispatch(documentdownloadaction({
                   "file": f.file.toString().substring(1, f.file.length)
@@ -128,9 +129,18 @@ function Employement() {
               }} style={{ cursor: "pointer", backgroundColor: "blue", padding: "4px", borderRadius: "4px", color: "white" }}> </i>
               <i className="pi pi-trash mr-2 ml-2" style={{ cursor: "pointer", backgroundColor: "red", padding: "4px", borderRadius: "4px", color: "white" }}
                 onClick={() => {
-                  dispatch(deletedocumentaction({
-                    "fileid": f.id
-                  }))
+                  confirmDialog({
+                    message: 'Do you want to delete this fi?',
+                    header: 'Delete Confirmation',
+                    icon: 'pi pi-info-circle',
+                    acceptClassName: 'p-button-danger',                                     
+                    accept: () => 
+                    dispatch(deletedocumentaction({
+                      "fileid": f.id
+                    })),
+                    reject: () => console.log()
+                })
+             
                 }}
               > </i>
 
@@ -172,11 +182,11 @@ function Employement() {
       <div className={className}>
 
         <span className={titleClassName}>
-          <h4>
+          <h5>
             Employement details
-          </h4>
+          </h5>
         </span>
-        {roles.includes("HR") && <SelectButton value={gridview} options={gridviewoptions} onChange={(e) => setgridview(e.value)} />}
+        {/* {roles.includes("HR") && <SelectButton value={gridview} options={gridviewoptions} onChange={(e) => setgridview(e.value)} />} */}
         <Button className='' style={{}} onClick={e => { setEditmode(false); setModaldialog(true) }}>Add </Button>            </div>
     )
   }
@@ -216,10 +226,20 @@ function Employement() {
 
 
         }} />
-        <Button style={{ height: "40px", width: "4.4rem" }} icon="pi pi-trash" className="p-button-danger" onClick={() => dispatch(deleteemployementdetailsaction({
-          "id": rowdata.id
+        <Button style={{ height: "40px", width: "4.4rem" }} icon="pi pi-trash" className="p-button-danger" onClick={() => 
+            confirmDialog({
+              message: 'Do you want to delete this record?',
+              header: 'Delete Confirmation',
+              icon: 'pi pi-info-circle',
+              acceptClassName: 'p-button-danger',                                     
+              accept: () => 
+              dispatch(deleteemployementdetailsaction({
+                "id": rowdata.id
+              })),
+              reject: () => console.log()
+          })
 
-        }))} />
+        } />
 
       </div>
 
@@ -261,17 +281,12 @@ function Employement() {
            
            }
           
-            .card1{
-                   
-                   
-                // background: linear-gradient(-52deg, #155dce 0%, #2de3b0 100%);
-                // background:  linear-gradient(-50deg, #155dce 0%, #2de3b0 100%);
-                background:  #c1c2f3;;
-                    background-blend-mode: normal;
-                    border-radius:8px;
-                    // width:300px
-                    
-               color:white;
+           .card1 {
+            background-color: var(--surface-card);
+            padding: 1.5rem;
+            margin-bottom: 1rem;
+            border-radius: 12px;
+            box-shadow: 0px 3px 5px #6366f1, 0px 0px 2px #6366f1, 0px 1px 4px #6366f1 !important;
             }
             .card1:hover{
                 scale:1.1;
@@ -314,6 +329,7 @@ function Employement() {
     padding:  0rem; 
 }
             `}</style>
+            <ConfirmDialog/>
       <Panel headerTemplate={template}>
 
         <Dialog  position='top' visible={modalDialog} style={{ width: "450px" }} header={editmode ? "Edit  Information " : "Add  Information "} modal className="p-fluid" onHide={() => setModaldialog(false)}>
@@ -534,19 +550,19 @@ function Employement() {
                                     </Button>
                                 </div>
                             </Dialog>
-              <div className="p-fluid  grid filesarea" style={{}}>
+              <div className="p-fluid  grid" style={{}}>
 
 
 
               {e.files.length > 0 ?
                                     e.files.map((f) => (
-                                        <div className='field col-12 md:col-12 flex' style={{ border: "2px solid blue", padding: "2px", marginTop: "4px", verticalAlign: "center", borderRadius: "10px", backgroundColor: "#C1C2F3", justifyContent: "space-between" }}>
+                                        <div className='field col-12 md:col-12 flex' style={{ border: "0px solid blue", color:"blue", padding: "2px", marginTop: "4px", verticalAlign: "center",  backgroundColor: "#f8f9fa", justifyContent: "space-between" }}>
                                             <div className="fileleftdiv">
-                                                {f.file.split("/")[f.file.split("/").length - 1].toString().length < 20 ? f.file.split("/")[f.file.split("/").length - 1] : f.file.split("/")[f.file.split("/").length - 1].substring(0, 25) + "..."}
+                                                {f.file.split("/")[f.file.split("/").length - 1].toString().length < 20 ? f.file.split("/")[f.file.split("/").length - 1] : f.file.split("/")[f.file.split("/").length - 1]}
                                                 <br />
                                                 {
-                                                    (f.verificationcomments == "" || f.verificationcomments == null) ? <></> : <>{
-                                                        "Comments : " + f.verificationcomments}</>
+                                                    (f.verificationcomments == "" || f.verificationcomments == null) ? <></> : <div className="pt-2" style={{color:"red"}}>{
+                                                        "Comments : " + f.verificationcomments}</div>
                                                 }
                                             </div>
                                             <div className='filerightdiv'>
@@ -734,18 +750,19 @@ function Employement() {
                                                 }} style={{ cursor: "pointer", backgroundColor: "blue", padding: "4px", borderRadius: "4px", color: "white", height: "25px" }}> </i>
                                                 {((!Object.is(f.verified, "pending") && !Object.is(f.verified, "verified")) || roles.includes("HR")) ? <i className="pi pi-trash mr-2 " style={{ cursor: "pointer", backgroundColor: "red", padding: "4px", borderRadius: "4px", color: "white", height: "25px" }}
                                                     onClick={() => {
+                                                      
                                                         confirmDialog({
-                                                            message: 'Are you sure you want to proceed?',
-                                                            header: 'Confirmation',
-                                                            icon: 'pi pi-exclamation-triangle',
-                                                            position: "top",
-                                                            accept: () => alert("success"),
-                                                            reject: () => alert("rej")
+                                                            message: 'Do you want to delete this file?',
+                                                            header: 'Delete Confirmation',
+                                                            icon: 'pi pi-info-circle',
+                                                            acceptClassName: 'p-button-danger',                                     
+                                                            accept: () => 
+                                                            dispatch(deletedocumentaction({
+                                                                "fileid": f.id
+                                                            })),
+                                                            reject: () => console.log()
                                                         });
 
-                                                        // dispatch(deletedocumentaction({
-                                                        //     "fileid": f.id
-                                                        // }))
 
 
 
@@ -755,7 +772,7 @@ function Employement() {
                                             </div>
                                         </div>
                                     )) :
-                                    <div className='nofiles' style={{}}
+                                    <div className='field col-12 md:col-12 flex nofiles' style={{ border: "0px solid blue", padding: "2px", marginTop: "4px", verticalAlign: "center", alignContent:"center", borderRadius: "10px", backgroundColor: "#f8f9fa", justifyContent: "space-around" }}
                                         onDragOver={e => {
                                             e.preventDefault();
 
@@ -778,14 +795,14 @@ function Employement() {
               </div>
               <br />
               <div className="card-footer p-fluid grid ">
+                <div className="field col-12 md:col-5 flex">
+
+                </div>
                 <div className="field col-12 md:col-3 flex">
 
-                </div>
-                <div className="field col-12 md:col-2 flex">
-
 
                 </div>
-                <div className="field col-12 md:col-7 flex gap-2">
+                <div className="field col-12 md:col-4 flex gap-2">
                   <Button className="p-button-info editbutton mr-2" style={{ height: "35px", width: "3.5rem" }} label="" icon="pi pi-pencil" onClick={() => { setEditmode(true); settempdata(e); setModaldialog(true); }}></Button>
 
 
@@ -819,10 +836,22 @@ function Employement() {
 
 
                   }} />
-                  <Button style={{ height: "35px", width: "3.5rem" }} icon="pi pi-trash" className="p-button-danger" onClick={() => dispatch(deleteemployementdetailsaction({
-                    "id": e.id
+                  <Button style={{ height: "35px", width: "3.5rem" }} icon="pi pi-trash" className="p-button-danger" onClick={() =>
+       
+                  confirmDialog({
+                    message: 'Do you want to delete this record?',
+                    header: 'Delete Confirmation',
+                    icon: 'pi pi-info-circle',
+                    acceptClassName: 'p-button-danger',                                     
+                    accept: () => 
+                    dispatch(deleteemployementdetailsaction({
+                        "id": e.id
 
-                  }))} />
+                    })),
+                    reject: () => console.log()
+                })
+
+                  } />
 
                 </div>
